requirements.txt
Django>=5.2.1
django-easy-audit>=1.3.7
djangorestframework>=3.16.0
graphene-django>=3.2.3
django-cors-headers>=4.7.0
geoip2>=5.1.0
psycopg2-binary>=2.9.9 